2021-10-26:update

+++ src/easy_task/module/task_for_event.py
@@ -263,24 +263,31 @@ class EventExtractionTask(SequenceTaggingTask):
                 self.save_checkpoint(cpt_file_name='{}.cpt.{}.{}.e{}.b{}.p{}.s{}'.format(\
                     self.setting.task_name, data_type, 0, self.setting.num_train_epochs, self.setting.train_batch_size, self.setting.percent, self.setting.seed))
                 
-            if self.setting.save_cpt_flag == 1:
+            save_cpt_file = '{}.cpt.{}.e({}).b({}).p({}).s({})'.format(\
+                    self.setting.task_name, epoch, self.setting.num_train_epochs, self.setting.train_batch_size, str(self.setting.percent).replace('.','|'), self.setting.seed)
+            if self.setting.save_cpt_flag == 1 and not os.path.exists(os.path.join(self.setting.model_dir, save_cpt_file)):
                 # save last epoch
                 last_epoch = self.get_latest_cpt_epoch()
                 if last_epoch != 0:
                     # delete lastest epoch model and store this epoch
-                    delete_cpt_file = '{}.cpt.{}'.format(self.setting.task_name, last_epoch)
+                    delete_cpt_file = '{}.cpt.{}.e({}).b({}).p({}).s({})'.format(\
+                        self.setting.task_name, last_epoch, self.setting.num_train_epochs, self.setting.train_batch_size, str(self.setting.percent).replace('.','|'), self.setting.seed)
+
                     if os.path.exists(os.path.join(self.setting.model_dir, delete_cpt_file)):
                         os.remove(os.path.join(self.setting.model_dir, delete_cpt_file))
+                        self.logger.info('remove model {}'.format(delete_cpt_file))
                     else:
                         self.logger.info("{} does not exist".format(delete_cpt_file), level=logging.WARNING)
+
                 self.logger.info('saving latest epoch model...')
-                self.save_checkpoint(cpt_file_name='{}.cpt.{}.{}.e{}.b{}.p{}.s{}'.format(\
-                    self.setting.task_name, data_type, epoch, self.setting.num_train_epochs, self.setting.train_batch_size, self.setting.percent, self.setting.seed))
-            elif self.setting.save_cpt_flag == 2:
+                self.save_checkpoint(cpt_file_name='{}.cpt.{}.e({}).b({}).p({}).s({})'.format(\
+                    self.setting.task_name, epoch, self.setting.num_train_epochs, self.setting.train_batch_size, str(self.setting.percent).replace('.','|'), self.setting.seed))
+
+            elif self.setting.save_cpt_flag == 2 and not os.path.exists(os.path.join(self.setting.model_dir, save_cpt_file)):
                 # save each epoch
                 self.logger.info('saving epoch {} model...'.format(epoch))
-                self.save_checkpoint(cpt_file_name='{}.cpt.{}.{}.e{}.b{}.p{}.s{}'.format(\
-                    self.setting.task_name, data_type, epoch, self.setting.num_train_epochs, self.setting.train_batch_size, self.setting.percent, self.setting.seed))
+                self.save_checkpoint(cpt_file_name='{}.cpt.{}.e({}).b({}).p({}).s({})'.format(\
+                    self.setting.task_name, epoch, self.setting.num_train_epochs, self.setting.train_batch_size, str(self.setting.percent).replace('.','|'), self.setting.seed))
 
 
     def custom_collate_fn_train(self, features: list) -> list:
